pyproject.toml
[project]
name = "vaultflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow==2.9.3",
    "awscli>=1.44.80",
    "boto3>=1.42.90",
    "dbt-postgres>=1.10.0",
    "faker>=40.13.0",
    "pandas>=3.0.2",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
]
